fileop.c: Important edit for parallel read of dump files
In the previous version of fileop.c, the header file was read
independently by each MPI process. This caused serious slowdown when
the number of processes ran into several thousands. The system
administrator sometimes stopped the job because of system overload.

In this new version of fileop.c, modifications have been made in the
functions fread_restartfile_mpi and fread_restartfile_mpi_hdf5 such
that only PROCID=0 opens the header file. Reading in of the big
datafile is done via MPI_File_open and MPI_File_read in the case of
binary data, and H5Dread in the case of hdf5 files. Reading in restart
dump files is now extremely fast.

Since the previous slowdown occurred because of too many fopen
operations, the new version prints out a line every time fopen is
encountered. For reading in dump files, the output in this new version
of fileop.c will show that fopen is encountered only once (previously
it happened once for each MPI process).